Life/requirements.txt
Pillow
SpeechRecognition
asyncpg
dblpy
discord.py
granitepy
jishaku
matplotlib
objgraph
psutil
youtube-dl
aiOsu